src/api/requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.23,<1
rapidfuzz>=3,<4
PyYAML>=6,<7
elasticsearch>=8,<9
prometheus-client>=0.20,<1
pytest>=7,<9
pytest-asyncio>=0.21,<1
jsonschema>=4,<5
httpx>=0.24,<1
scrapy>=2.11,<3

